Revamp how error processing works, increasing user control and resiliancy.
Instead of a variable to control whether we debug or not, we now have
a on-error callback that gets called on every error in a callback.

Also, now all frame processing is done at once, after all incoming
bytes have been processed.  This will decrease the chance of
disruptions because of errors.  However, there can be no safe error
when it is thrown, since if several frames are processed, and the
first one throws an error, there is no mechanism to go back and
process the rest.