Fix incorrect re-throwing of error condition.
This was passing a cons cell as the first argument to `error', which
expects a format string and thus throws an error itself.  Use the
re-throw idiom suggested by
http://www.gnu.org/software/emacs/manual/html_node/elisp/Handling-Errors.html#index-condition_002dcase-578
instead.

NB: at first glance, `(error (cdr err))' is preferable as it is more
succinct.  However, the generic form will keep working should the
CONDITION-NAME in the condition-case form ever be changed.