execute handshake after process status moved to open
if nowait is t, `open-network-stream` or `make-network-process`
function returns without waiting connection opened.
we need to wait executing handshake process until process status
changed to `open`.
we call `websocket-handshake` in sentinel because sentinel function is
executed when process status changed to `open`.

if nowait is nil, `open-network-stream` or `make-network-process`
returns after connection opened.
sentinel function is not executed when process-status changed to
`open`, we have to call `websocket-handshake` in `websocket-open`.