Modernize & rewrite functional tests as ert tests.
Between the wss echo server, and the websocket library's ability to be a server
and a client, there isn't as much need for the Tornado server anymore.  This
commit removes the server, and the tests that needed it.

The existing tests are rewritten as ert tests, so that they can be run with the
ert framework, which is easier and more standardized than the previous method of
running the functional tests on the command line.

Without the Tornado server, there is substantially less exernal depencies, but
not none, so the comments now warn about various causes of failures.